refactor(ios): fold the talk button's transition into its only caller
`TalkButtonInteraction.swift` sat in `LukeKit`, shared with nothing: the
watch's hold-to-talk screen reads its own gesture, and the phone's
`VoiceView` was the file's one caller. It moves next to the gesture that
reads it, and its three cases follow into the app target's own suite, which
the testable app target now makes possible.
