+++ apps/ios/LukeKit/Tests/LukeKitTests/RealtimeSessionTests.swift
@@ -176,32 +176,6 @@ final class RealtimeWebSocketAuthenticationTests: XCTestCase {
     }
 }
 
-final class TalkButtonInteractionTests: XCTestCase {
-    func testQuickFirstTapLatchesListening() {
-        XCTAssertEqual(
-            talkButtonReleaseAction(heldDuration: 0.04, wasLatched: false),
-            .latch
-        )
-    }
-
-    func testHoldSendsOnFirstRelease() {
-        XCTAssertEqual(
-            talkButtonReleaseAction(
-                heldDuration: talkButtonTapDuration + 0.001,
-                wasLatched: false
-            ),
-            .send
-        )
-    }
-
-    func testSecondTapSendsLatchedTurn() {
-        XCTAssertEqual(
-            talkButtonReleaseAction(heldDuration: 0.01, wasLatched: true),
-            .send
-        )
-    }
-}
-
 // MARK: - State machine tests
 
 @MainActor
